- problems with login and sign up solved
