+++ Notes/DataManager/Persistence.swift
@@ -77,140 +77,4 @@ struct PersistenceController {
 			}
 		}
 	}
-	
-	func createNote(title: String, timestamp: Date, id: Int32, user: User) -> Note {
-		let note = Note(context: container.viewContext)
-		note.id = id
-		note.timestamp = timestamp
-		note.title = title
-		note.user = user
-		user.addToNote(note)
-		save(container: container)
-		return note
-	}
-	
-	func fetchAllNotes() -> [Note] {
-		let request: NSFetchRequest<Note> = Note.fetchRequest()
-		var fetchedNotes: [Note] = []
-		do {
-			fetchedNotes = try container.viewContext.fetch(request)
-		} catch let error {
-			print("Error fetching singers \(error)")
-		}
-		return fetchedNotes
-	}
-	
-	func fetchNotesByUser(user: User) -> [Note] {
-		let request: NSFetchRequest<Note> = Note.fetchRequest()
-		request.predicate = NSPredicate(format: "user = %@", user)
-		request.sortDescriptors = [NSSortDescriptor(key: "timestamp", ascending: false)]
-		var fetchedNotes: [Note] = []
-		do {
-			fetchedNotes = try container.viewContext.fetch(request)
-		} catch let error {
-			print("Error fetching notes \(error)")
-		}
-		return fetchedNotes
-	}
-	
-	func fetchNotesById(id: Int32) -> Note? {
-		let request: NSFetchRequest<Note> = Note.fetchRequest()
-		request.predicate = NSPredicate(format: "id = %@", id as Int32)
-		request.sortDescriptors = [NSSortDescriptor(key: "timestamp", ascending: false)]
-		var fetchedNote: Note?
-		do {
-			fetchedNote = try container.viewContext.fetch(request).first
-		} catch let error {
-			print("Error fetching notes \(error)")
-		}
-		return fetchedNote
-	}
-	
-	func updateNote(title: String, timestamp: Date, id: Int32, data: Data, user: User) -> Note {
-		let context = container.viewContext
-		let note: Note!
-		
-		let fetchNote: NSFetchRequest<Note> = Note.fetchRequest()
-		fetchNote.predicate = NSPredicate(format: "id = %i", id)
-		
-		let results = try? context.fetch(fetchNote)
-		
-		if results?.count == 0 {
-			// here you are inserting
-			note = Note(context: context)
-		} else {
-			// here you are updating
-			note = (results?.first)
-		}
-		note.id = id
-		note.timestamp = timestamp
-		note.title = title
-		note.noteData = data
-		note.user = user
-		user.addToNote(note)
-		
-		save(container: container)
-
-		return note
-	}
-	
-	func deleteNote(note: Note) {
-		let context = container.viewContext
-		context.delete(note)
-		save(container: container)
-	}
-	
-	func createUser(username: String, email: String, password: String, id: Int32) {
-		let user = User(context: container.viewContext)
-		user.username = username
-		user.email = email
-		user.id = id
-		user.userId = UUID()
-		user.password = password
-		save(container: container)
-	}
-	
-	func fetchAllUsers() -> [User] {
-		let request: NSFetchRequest<User> = User.fetchRequest()
-		var fetchedUsers: [User] = []
-		do {
-			fetchedUsers = try container.viewContext.fetch(request)
-		} catch let error {
-			print("Error fetching singers \(error)")
-		}
-		return fetchedUsers
-	}
-	
-	func fetchUsersByUsername(username: String) -> User? {
-		let request: NSFetchRequest<User> = User.fetchRequest()
-		request.predicate = NSPredicate(format: "username = %@", username)
-		var fetchedUser: User?
-		do {
-			fetchedUser = try container.viewContext.fetch(request).first
-		} catch let error {
-			print("Error fetching users \(error)")
-		}
-		return fetchedUser
-	}
-	
-	func fetchUsersByUsernameAndPassword(username: String, password: String) -> User? {
-		let request: NSFetchRequest<User> = User.fetchRequest()
-		let predicate1 = NSPredicate(format: "username = %@", username)
-		let predicate2 = NSPredicate(format: "password = %@", password)
-		let compound = NSCompoundPredicate(andPredicateWithSubpredicates: [predicate1, predicate2])
-		request.predicate = compound
-		var fetchedUser: User?
-		do {
-			fetchedUser = try container.viewContext.fetch(request).first
-		} catch let error {
-			print("Error fetching users \(error)")
-		}
-		return fetchedUser
-	}
-	
-	func deleteUser(user: User) {
-		let context = container.viewContext
-		context.delete(user)
-		save(container: container)
-	}
 }
