+++ Notes/DataManager/UserDataManager.swift
@@ -9,33 +9,12 @@ import Foundation
 import CoreData
 
 struct UserDataManager {
-	let container: NSPersistentContainer
+	private let container: NSPersistentContainer
 	private var persistenceController : PersistenceController
 	
-	init(inMemory: Bool = false, persistenceController: PersistenceController) {
+	init(container: NSPersistentContainer, persistenceController: PersistenceController) {
+		self.container = container
 		self.persistenceController = persistenceController
-		
-		container = NSPersistentContainer(name: "Notes")
-		if inMemory {
-			container.persistentStoreDescriptions.first!.url = URL(fileURLWithPath: "/dev/null")
-		}
-		container.loadPersistentStores(completionHandler: { (storeDescription, error) in
-			if let error = error as NSError? {
-				// Replace this implementation with code to handle the error appropriately.
-				// fatalError() causes the application to generate a crash log and terminate. You should not use this function in a shipping application, although it may be useful during development.
-				
-				/*
-				 Typical reasons for an error here include:
-				 * The parent directory does not exist, cannot be created, or disallows writing.
-				 * The persistent store is not accessible, due to permissions or data protection when the device is locked.
-				 * The device is out of space.
-				 * The store could not be migrated to the current model version.
-				 Check the error message to determine what the actual problem was.
-				 */
-				fatalError("Unresolved error \(error), \(error.userInfo)")
-			}
-		})
-		container.viewContext.automaticallyMergesChangesFromParent = true
 	}
 	
 	func createUser(username: String, email: String, password: String, id: Int32) {
